latest commit

+++ database/migrations/2022_07_23_073102_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->bigInteger('price');
+            $table->decimal('price', 8, 2);
             $table->string('quantity');
             $table->timestamps();
         });
